+++ astropy/tests/plugins/__init__.py
@@ -1,0 +1,4 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This package contains pytest plugins that are used by the astropy test suite.
+"""
