+++ astropy/tests/plugins/openfiles.py
