+++ astropy/tests/plugins/remotedata.py
@@ -6,8 +6,8 @@ into conftest.py in the root directory.
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-from .helper import pytest
-from .disable_internet import turn_off_internet, turn_on_internet
+from ..helper import pytest
+from ..disable_internet import turn_off_internet, turn_on_internet
 
 
 def pytest_addoption(parser):
