+++ astropy/tests/plugins/repeat.py
@@ -6,7 +6,7 @@ into conftest.py in the root directory.
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-from ..extern.six.moves import range
+from ...extern.six.moves import range
 
 
 def pytest_addoption(parser):
